+++ changelog.md
@@ -16,6 +16,8 @@ The Release-Candidate
 - Added Team-Switch option for Round Two or always
 - Bots stand still during spawn-protection
 - Lots of refactoring
+- new paths for Karkhand
+- longer reaction-time on bigger distances
 
 
 ## [Unreleased - 1.7.1.X] - 2021-05-XX